Update SupportBrowserHistory.php
The `sortKeys()` method call in the collection pipeline within the `mergeComponentPropertiesWithExistingQueryParamsFromOtherComponentsAndTheRequest()` method causes query strings to be re-written in alphabetical order on any page with multiple query strings in the URL.

This should not be default behavior. Sites that rely upon query strings for locating separate resources can receive duplicate content SEO warnings from this behavior. We have confirmed this side effect in production on a high traffic message board.