Account for Varying Emission Params
In our project we have an enum with the value "info" which is one of the parameters for one of our Livewire emitters.

This function thinks that "info" is a callable because PHP has a built-in function `info()` and therefore our tests for it fail.

A simple `! is_string()` check should suffice!