Suppress type warning (#4376)
I'm not sure that this is the best solution, but since `parse_str()` expects a string, it made sense to me to cast whatever **this particular call to** `parse_url()` returns to a string.

It seems that `parse_url()`, when specifying the `PHP_URL_QUERY` component, will always return a string or `null`.  So, this seems like a safe solution.

FYI...this is only generating a warning:
```bash
parse_str(): Passing null to parameter #1 ($string) of type string is deprecated in /path/to/vendor/livewire/livewire/src/Features/SupportBrowserHistory.php on line 130
```

So, it's not fixing anything that's broken, but just silencing some of the noise. 🤓 

Please let me know if anyone has any questions/concerns.