Update teleport.md
I noticed that in order to have a component that is reactive and have a `@teleport()` directive on the page it must be wrapped in a `wire:ignore`. If this is a bug, I am happy to open an issue but I didn't want others to be stuck on this like I was.