Allow download of preview files directly from storage
This commit introduces support for non-local storages again. This time
the cache hit path is fixed and tested which caused #7794 before.

Also the cachedFileResponse function has been changed to de-duplicate
the matchesCache check and to not return charset applied to the
Content-Type.

To also support sending the header in the test, the If-Modified-Since
header is now fetched from the Laravel Request object, rather than from
$_SERVER.