fix NProgress not complete on page load
issue:
When using Livewire with NProgress, the progress bar starts when navigating, but it does not complete when loading a new page. It remains visible until it's manually hidden or replaced, making it seem incomplete.

Solution:

I fixed this by adding an event listener for livewire:navigated to properly trigger NProgress.done() once the page has finished loading. This ensures the progress bar completes and disappears as expected after navigation.